+++ src/main/kotlin/uk/co/ben_gibson/git/link/git/RepositoryExtensions.kt
@@ -3,5 +3,5 @@ package uk.co.ben_gibson.git.link.git
 import git4idea.GitUtil
 import git4idea.repo.GitRepository
 
-fun GitRepository.findRemote(name: String) = GitUtil.findRemoteByName(this, name)
+fun GitRepository.locateRemote(name: String) = GitUtil.findRemoteByName(this, name)
 fun GitRepository.currentCommit() = currentRevision?.let { Commit(it) }