+++ src/main/kotlin/uk/co/ben_gibson/git/link/pipeline/RecordHitMiddleware.kt
@@ -2,16 +2,14 @@ package uk.co.ben_gibson.git.link.pipeline
 
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
-import com.intellij.openapi.project.Project
 import uk.co.ben_gibson.git.link.settings.ApplicationSettings
 import java.net.URL
-import uk.co.ben_gibson.git.link.Context
 
 @Service
 class RecordHitMiddleware : Middleware {
-    override val priority = 2
+    override val priority = 20
 
-    override fun invoke(project: Project, context: Context, next: () -> URL?) : URL? {
+    override fun invoke(pass: Pass, next: () -> URL?) : URL? {
         val url = next() ?: return null
 
         service<ApplicationSettings>().incrementHits();