+++ src/main/kotlin/uk/co/ben_gibson/git/link/url/UrlOptions.kt
@@ -5,24 +5,31 @@ import uk.co.ben_gibson.git.link.git.File
 import uk.co.ben_gibson.git.link.ui.LineSelection
 import java.net.URL
 
-interface LineSelectionAware {
-    val lineSelection: LineSelection?
-}
-
 sealed class UrlOptions(val baseUrl: URL)
 
 class UrlOptionsCommit(baseUrl: URL, val commit: Commit) : UrlOptions(baseUrl)
 
+interface UrlOptionsFileAware {
+    val baseUrl: URL
+    val file: File
+    val ref: String
+    val lineSelection: LineSelection?
+}
+
 class UrlOptionsFileAtCommit(
     baseUrl: URL,
-    val file: File,
+    override val file: File,
     val commit: Commit,
     override val lineSelection: LineSelection? = null
-) : UrlOptions(baseUrl), LineSelectionAware
+) : UrlOptions(baseUrl), UrlOptionsFileAware {
+    override val ref: String get() = commit.toString()
+}
 
 class UrlOptionsFileAtBranch(
     baseUrl: URL,
-    val file: File,
+    override val file: File,
     val branch: String,
     override val lineSelection: LineSelection? = null
-) : UrlOptions(baseUrl), LineSelectionAware
+) : UrlOptions(baseUrl), UrlOptionsFileAware {
+    override val ref: String get() = branch
+}